@@ -7,7 +7,7 @@ name = "proxy-scraper-checker"
 version = "0.0.0"
 license = { text = "MIT" }
 authors = [{ name = "monosans", email = "hsyqixco@protonmail.com" }]
-requires-python = ">=3.8,<4"
+requires-python = ">=3.9,<4"
 classifiers = ["Private :: Do Not Upload"]
 dependencies = [
   "aiofiles",
@@ -26,7 +26,7 @@ dependencies = [
 ]
 optional-dependencies.non-termux = [
   "aiodns>=1.1,<4",
-  "uvloop>=0.14,<0.21; implementation_name=='cpython' and (sys_platform=='darwin' or sys_platform=='linux')",
+  "uvloop>=0.14,<0.22; implementation_name=='cpython' and (sys_platform=='darwin' or sys_platform=='linux')",
 ]
 optional-dependencies.nuitka = ["nuitka==2.4.10"]
 urls.repository = "https://github.com/monosans/proxy-scraper-checker"
