@@ -57,17 +57,20 @@ class Proxy:
                 password=self.password,
                 ssl=SSL_CONTEXT,
             )
-            async with ClientSession(
-                connector=connector,
-                headers=HEADERS,
-                cookie_jar=get_cookie_jar(),
-                raise_for_status=True,
-                timeout=settings.timeout,
-                fallback_charset_resolver=fallback_charset_resolver,
-            ) as session, session.get(
-                settings.check_website,
-                headers=settings.check_website_type.headers,
-            ) as response:
+            async with (
+                ClientSession(
+                    connector=connector,
+                    headers=HEADERS,
+                    cookie_jar=get_cookie_jar(),
+                    raise_for_status=True,
+                    timeout=settings.timeout,
+                    fallback_charset_resolver=fallback_charset_resolver,
+                ) as session,
+                session.get(
+                    settings.check_website,
+                    headers=settings.check_website_type.headers,
+                ) as response,
+            ):
                 content = await response.read()
         self.timeout = perf_counter() - start
         if settings.check_website_type == CheckWebsiteType.HTTPBIN_IP:
